+++ src/vector_press/agent/__init__.py
@@ -5,7 +5,7 @@ from src.vector_press.agent.agent import VectorPressAgent, AgentState, _should_c
 from .news_api_client import GuardianAPIClient, BaseNewsAPIClient, _extract_article_text
 from .rss_client import TechnologyRSSClient, SportsRSSClient
 # Validation models
-from .tools_validation import TavilySearch, TheGuardianApi, NewYorkTimesApi, TechnologyRSSFeed, SportsRSSFeed
+from .tools import TavilySearch, TheGuardianApi, NewYorkTimesApi, TechnologyRSSFeed, SportsRSSFeed
 
 # Export all public classes and functions
 __all__ = [
