+++ src/vector_press/agent/news_api_client.py
@@ -4,7 +4,7 @@ from typing import Dict
 #from datetime import datetime
 import requests
 
-from src.vector_press.agent.tools_validation import (TheGuardianApi, NewYorkTimesApi)
+from src.vector_press.agent.tools import (TheGuardianApi, NewYorkTimesApi)
 
 from config import settings
